common: captures the state of splitting C99 and C++ interfaces for the context submodule
