+++ sixtracklib/common/context/CMakeLists.txt
@@ -1,5 +1,7 @@
 set( SIXTRL_COMMON_CONTEXT_C99_HEADERS
      compute_arch.h
+     context_base.h
+     argument_base.h
 )
 
 set( SIXTRL_COMMON_CONTEXT_C99_SOURCES
@@ -27,16 +29,18 @@ target_compile_options( sixtrack_common_context PRIVATE
 
 set( SIXTRL_COMMON_CONTEXT_CXX_HEADERS
      definitions.h
-     context_base.h
-     context_base_with_nodes.h
-     argument_base.h
+     context_base.hpp
+     context_base_with_nodes.hpp
+     argument_base.hpp
 )
 
 set( SIXTRL_COMMON_CONTEXT_CXX_SOURCES
      context_base.cpp
+     context_base_c99.cpp
      context_base_with_nodes.cpp
-     context.cpp
+     context_base_with_nodes_c99.cpp
      argument_base.cpp
+     argument_base_c99.cpp
 )
 
 add_library( sixtrack_common_context_cxx OBJECT
